+++ src/components/shared/activity-card/issue-card.tsx
@@ -5,7 +5,7 @@ import { formatDistanceToNow } from "date-fns";
 import { Contributor } from "@/components/shared/contributor-avatar-group";
 import { cn } from "@/lib/utils";
 
-import ActivityCardBase from "./shared/ActivityCardBase";
+import ActivityCardBase from "./shared/activity-card-base";
 
 export type Props = {
   title: string;
