+++ src/components/shared/activity-card/shared/activity-card-base.tsx
